Update PerfectNumber.java
After thinking for days whether I needed to implement an interface, a class or an enumerator (of which an enumerator seemed the most logical), I finally found my answer after spending fifteen seconds on the "More on lambda expressions & streams" source. The second I saw an enumerator was being implemented I felt like I had seen the light. Finally, the program actually seems rational.

Minor to moderate note: one part of the code doesn't match the requirements of functional programming (as far as I'm aware), which is the implementation of if-if else-else (essentially a tri-case switch, see lines 30-32). I did try to find a solution for implementing branching /functionally/, but every solution seemed way too convoluted to be logical to use in a first assignment.