Update PerfectNumber.java
Replaced the if-if else-else statement with a hashMap. However, the map's functionality is incredibly, well, thick for lack of a better word. A lot of lines of code for very small functionality. Immediate next goal is to see if there is an achievable way to shorten the construction of the map. Frankly, I think the map as it is is incredibly obnoxious. It's so ugly, look at it.